sys/log: Fix some mistakes in log API fn decls
The printf functions accepted 16-bit modules and levels.  These values
are defined to be 8-bit and are persisted as 8-bit values.  Change the
parameter types from `uint16_t` to `uint8_t`.

Also change the `msg` paremeter in `sys/log/full` from `char *` to
`const char *`.

+++ sys/log/modlog/include/modlog/modlog.h
@@ -201,7 +201,7 @@ int modlog_foreach(modlog_foreach_fn *fn, void *arg);
  * @param level                 The severity of the log entry to write.
  * @param msg                   The "printf" formatted string to write.
  */
-void modlog_printf(uint16_t module, uint16_t level, const char *msg, ...);
+void modlog_printf(uint8_t module, uint8_t level, const char *msg, ...);
 
 #else /* LOG_FULL */
